optimized for dataset resumption.
